Start removing flyway and moving shared test utils to new project
move common test utils to common location, resolve issues

Clean up poms, fix RDA migration tests, moving some things to common loations

Remove schema migration from RDA load

Moving things to common locations and fixing tests

Fix issues with bfd IT tests due to not writing out the server file

Cleanup

delete moved files after rebase

fix rda tests

minor adjustments to try and help clean schema between tests

remove table drop

Fix rda tests by dropping type after their migrations run

Fix javadocs, add cleanup for IT tests

Dont move rda tests

add comment

add drop type for json for new script

add debugging to failure of data set processing due to test failure

Add migration to pipeline datasource on setup

Add ability to create unmigrated datasource and add migration to datasource creation for tests automatically

Add cleanup for pipeline IT tests

Moving the schema cleanup into the schema cleanup method in database utils and out of the migrator tests
