+++ apps/bfd-db-migrator/pom.xml
@@ -14,7 +14,6 @@
 	<description>An application integrated with Flyway and Hibernate responsible for executing database migrations and validations against BFD datastores.</description>
 
 	<properties>
-		<flyway.version>8.5.13</flyway.version>
 		<reflections.version>0.10.2</reflections.version>
 	</properties>
 
@@ -96,13 +95,6 @@
 			<version>${project.version}</version>
 		</dependency>
 		<dependency>
-			<!-- Provides a database schema mechanism. See the DatabaseSchemaManager
-				class. -->
-			<groupId>org.flywaydb</groupId>
-			<artifactId>flyway-core</artifactId>
-			<version>${flyway.version}</version>
-		</dependency>
-		<dependency>
 			<!-- Used to provide DB connection pooling. See https://github.com/brettwooldridge/HikariCP
                     for details. -->
 			<groupId>com.zaxxer</groupId>
@@ -160,6 +152,11 @@
 			<artifactId>mockito-core</artifactId>
 			<version>${mockito.version}</version>
 			<scope>test</scope>
+		</dependency>
+		<dependency>
+			<groupId>gov.cms.bfd</groupId>
+			<artifactId>bfd-shared-test-utils</artifactId>
+			<version>${project.version}</version>
 		</dependency>
 	</dependencies>
