+++ apps/bfd-db-migrator/src/main/java/gov/cms/bfd/migrator/app/AppConfiguration.java
@@ -1,80 +1,13 @@
 package gov.cms.bfd.migrator.app;
 
-import java.net.InetAddress;
-import java.net.UnknownHostException;
+import gov.cms.bfd.sharedutils.config.AppConfigurationException;
+import gov.cms.bfd.sharedutils.config.BaseAppConfiguration;
+import gov.cms.bfd.sharedutils.config.MetricOptions;
+import gov.cms.bfd.sharedutils.database.DatabaseOptions;
 import java.util.Optional;
 
-/**
- * Models the configuration options for the application. TODO: BFD-1558 Move this class into a
- * common location to be used here and pipeline, refactor getters
- */
-public final class AppConfiguration {
-
-  /**
-   * The name of the environment variable that should be used to provide the {@link
-   * #getDatabaseOptions()} {@link DatabaseOptions#getDatabaseUrl()} value.
-   */
-  public static final String ENV_VAR_KEY_DATABASE_URL = "DATABASE_URL";
-
-  /**
-   * The name of the environment variable that should be used to provide the {@link
-   * #getDatabaseOptions()} {@link DatabaseOptions#getDatabaseUsername()} value.
-   */
-  public static final String ENV_VAR_KEY_DATABASE_USERNAME = "DATABASE_USERNAME";
-
-  /**
-   * The name of the environment variable that should be used to provide the {@link
-   * #getDatabaseOptions()} {@link DatabaseOptions#getDatabasePassword()} value.
-   */
-  public static final String ENV_VAR_KEY_DATABASE_PASSWORD = "DATABASE_PASSWORD";
-
-  /**
-   * The name of the environment variable that should be used to provide the {@link
-   * #getDatabaseOptions()} {@link DatabaseOptions#getMaxPoolSize()} value.
-   */
-  public static final String ENV_VAR_KEY_DATABASE_MAX_POOL_SIZE = "DATABASE_MAX_POOL_SIZE";
-
-  /**
-   * The name of the environment variable that should be used to provide the {@link
-   * #getMetricOptions()} {@link MetricOptions#getNewRelicMetricKey()} value.
-   */
-  public static final String ENV_VAR_NEW_RELIC_METRIC_KEY = "NEW_RELIC_METRIC_KEY";
-
-  /**
-   * The name of the environment variable that should be used to provide the {@link
-   * #getMetricOptions()} {@link MetricOptions#getNewRelicAppName()} value.
-   */
-  public static final String ENV_VAR_NEW_RELIC_APP_NAME = "NEW_RELIC_APP_NAME";
-
-  /**
-   * The name of the environment variable that should be used to provide the {@link
-   * #getMetricOptions()} {@link MetricOptions#getNewRelicMetricHost()} value.
-   */
-  public static final String ENV_VAR_NEW_RELIC_METRIC_HOST = "NEW_RELIC_METRIC_HOST";
-
-  /**
-   * The name of the environment variable that should be used to provide the {@link
-   * #getMetricOptions()} {@link MetricOptions#getNewRelicMetricPath()} value.
-   */
-  public static final String ENV_VAR_NEW_RELIC_METRIC_PATH = "NEW_RELIC_METRIC_PATH";
-
-  /**
-   * The name of the environment variable that should be used to provide the {@link
-   * #getMetricOptions()} {@link MetricOptions#getNewRelicMetricPeriod()} value.
-   */
-  public static final String ENV_VAR_NEW_RELIC_METRIC_PERIOD = "NEW_RELIC_METRIC_PERIOD";
-
-  /**
-   * The name of the environment variable that should be used to provide the location of the flyway
-   * scripts.
-   */
-  public static final String ENV_VAR_FLYWAY_SCRIPT_LOCATION = "FLYWAY_SCRIPT_LOCATION";
-
-  /** Contains options for metrics capturing. */
-  private final MetricOptions metricOptions;
-
-  /** Contains options for database related configuration. */
-  private final DatabaseOptions databaseOptions;
+/** Models the configuration options for the application. */
+public class AppConfiguration extends BaseAppConfiguration {
 
   /**
    * Controls where flyway looks for migration scripts. If not set (null or empty string) flyway
@@ -96,27 +29,8 @@ public final class AppConfiguration {
       MetricOptions metricOptions,
       DatabaseOptions databaseOptions,
       String flywayScriptLocationOverride) {
-    this.metricOptions = metricOptions;
-    this.databaseOptions = databaseOptions;
+    super(metricOptions, databaseOptions);
     this.flywayScriptLocationOverride = flywayScriptLocationOverride;
-  }
-
-  /**
-   * Gets the {@link #metricOptions}.
-   *
-   * @return the {@link MetricOptions} that the application will use
-   */
-  public MetricOptions getMetricOptions() {
-    return metricOptions;
-  }
-
-  /**
-   * Gets the {@link #databaseOptions}.
-   *
-   * @return the {@link DatabaseOptions} that the application will use
-   */
-  public DatabaseOptions getDatabaseOptions() {
-    return databaseOptions;
   }
 
   /**
@@ -131,11 +45,7 @@ public final class AppConfiguration {
   /** {@inheritDoc} */
   @Override
   public String toString() {
-    StringBuilder builder = new StringBuilder();
-    builder.append(", databaseOptions=");
-    builder.append(databaseOptions);
-    builder.append(", metricsOptions=");
-    builder.append(metricOptions);
+    StringBuilder builder = new StringBuilder(super.toString());
     builder.append(", flywayScriptLocationOverride=");
     builder.append(flywayScriptLocationOverride);
     return builder.toString();
@@ -151,115 +61,15 @@ public final class AppConfiguration {
    * @throws AppConfigurationException An {@link AppConfigurationException} will be thrown if the
    *     configuration passed to the application are incomplete or incorrect.
    */
-  static AppConfiguration readConfigFromEnvironmentVariables() {
-    String databaseUrl = readEnvStringRequired(ENV_VAR_KEY_DATABASE_URL);
-    String databaseUsername = readEnvStringRequired(ENV_VAR_KEY_DATABASE_USERNAME);
-    String databasePassword = readEnvStringRequired(ENV_VAR_KEY_DATABASE_PASSWORD);
-    Optional<String> newRelicMetricKey = readEnvStringOptional(ENV_VAR_NEW_RELIC_METRIC_KEY);
-    Optional<String> newRelicAppName = readEnvStringOptional(ENV_VAR_NEW_RELIC_APP_NAME);
-    Optional<String> newRelicMetricHost = readEnvStringOptional(ENV_VAR_NEW_RELIC_METRIC_HOST);
-    Optional<String> newRelicMetricPath = readEnvStringOptional(ENV_VAR_NEW_RELIC_METRIC_PATH);
-    Optional<Integer> newRelicMetricPeriod = readEnvIntOptional(ENV_VAR_NEW_RELIC_METRIC_PERIOD);
-    Optional<Integer> databaseMaxPoolSize = readEnvIntOptional(ENV_VAR_KEY_DATABASE_MAX_POOL_SIZE);
+  public static AppConfiguration readConfigFromEnvironmentVariables() {
+
+    MetricOptions metricOptions = readMetricOptionsFromEnvironmentVariables();
+    DatabaseOptions databaseOptions = readDatabaseOptionsFromEnvironmentVariables();
+
     Optional<String> flywayScriptLocationOverride =
         readEnvStringOptional(ENV_VAR_FLYWAY_SCRIPT_LOCATION);
-    if (databaseMaxPoolSize.isPresent() && databaseMaxPoolSize.get() < 1) {
-      throw new AppConfigurationException(
-          String.format(
-              "Invalid value for configuration environment variable '%s': '%s'",
-              ENV_VAR_KEY_DATABASE_MAX_POOL_SIZE, databaseMaxPoolSize));
-    }
-
-    Optional<String> hostname;
-    try {
-      hostname = Optional.of(InetAddress.getLocalHost().getHostName());
-    } catch (UnknownHostException e) {
-      hostname = Optional.empty();
-    }
-
     String flywayScriptLocation = flywayScriptLocationOverride.orElse("");
 
-    MetricOptions metricOptions =
-        new MetricOptions(
-            newRelicMetricKey,
-            newRelicAppName,
-            newRelicMetricHost,
-            newRelicMetricPath,
-            newRelicMetricPeriod,
-            hostname);
-    DatabaseOptions databaseOptions =
-        new DatabaseOptions(
-            databaseUrl, databaseUsername, databasePassword, databaseMaxPoolSize.orElse(1));
-
     return new AppConfiguration(metricOptions, databaseOptions, flywayScriptLocation);
-  }
-
-  /**
-   * Reads an optional {@link String} from the environment variables by name. If not set, returns an
-   * empty {@link Optional}.
-   *
-   * @param environmentVariableName the name of the environment variable to get the value of
-   * @return the value of the specified environment variable, or {@link Optional#empty()} if it is
-   *     not set
-   */
-  static Optional<String> readEnvStringOptional(String environmentVariableName) {
-    Optional<String> environmentVariableValue =
-        Optional.ofNullable(System.getenv(environmentVariableName));
-    return environmentVariableValue;
-  }
-
-  /**
-   * Reads a {@link String} from the environment variables by name which is expected to exist. If
-   * the environment variable does not exist, throws an {@link AppConfigurationException}.
-   *
-   * @param environmentVariableName the name of the environment variable to get the value of
-   * @return the value of the specified environment variable
-   * @throws AppConfigurationException An {@link AppConfigurationException} will be thrown if the
-   *     value is missing.
-   */
-  static String readEnvStringRequired(String environmentVariableName) {
-    Optional<String> environmentVariableValue =
-        Optional.ofNullable(System.getenv(environmentVariableName));
-    if (!environmentVariableValue.isPresent()) {
-      throw new AppConfigurationException(
-          String.format(
-              "Missing value for configuration environment variable '%s'.",
-              environmentVariableName));
-    } else if (environmentVariableValue.get().isEmpty()) {
-      throw new AppConfigurationException(
-          String.format(
-              "Invalid value for configuration environment variable '%s': '%s'",
-              environmentVariableName, environmentVariableValue.get()));
-    }
-
-    return environmentVariableValue.get();
-  }
-
-  /**
-   * Reads an optional {@link Integer} from the environment variables by name. If not set, returns
-   * an empty {@link Optional}.
-   *
-   * @param environmentVariableName the name of the environment variable to get the value of
-   * @return the value of the specified environment variable, or {@link Optional#empty()} if it is
-   *     not set
-   * @throws AppConfigurationException An {@link AppConfigurationException} will be thrown if the
-   *     value cannot be parsed.
-   */
-  static Optional<Integer> readEnvIntOptional(String environmentVariableName) {
-    Optional<String> environmentVariableValueText =
-        Optional.ofNullable(System.getenv(environmentVariableName));
-    if (!environmentVariableValueText.isPresent()) {
-      return Optional.empty();
-    }
-
-    try {
-      return Optional.of(Integer.valueOf(environmentVariableValueText.get()));
-    } catch (NumberFormatException e) {
-      throw new AppConfigurationException(
-          String.format(
-              "Invalid value for configuration environment variable '%s': '%s' (%s)",
-              environmentVariableName, environmentVariableValueText.get(), e.getMessage()),
-          e);
-    }
   }
 }
