+++ apps/bfd-db-migrator/src/test/resources/db/migration-test/error-scenarios/bad-sql/V1__Create_beneficiary_table.sql
