+++ apps/bfd-db-migrator/src/test/resources/db/migration-test/error-scenarios/bad-sql/V2__Add_column.sql
