+++ apps/bfd-db-migrator/src/test/resources/db/migration-test/error-scenarios/bad-sql/V3__Add_bad_constraint.sql
