+++ apps/bfd-db-migrator/src/test/resources/db/migration-test/error-scenarios/bad-sql/V4__Add_column_again.sql
