+++ apps/bfd-db-migrator/src/test/resources/db/migration-test/error-scenarios/duplicate-version/V1__Create_beneficiary_table.sql
