+++ apps/bfd-db-migrator/src/test/resources/db/migration-test/error-scenarios/duplicate-version/V2__Add_column.sql
