+++ apps/bfd-db-migrator/src/test/resources/db/migration-test/error-scenarios/duplicate-version/V3__Alter_column_again.sql
