+++ apps/bfd-db-migrator/src/test/resources/db/migration-test/error-scenarios/duplicate-version/V3__Another_v3_script.sql
