+++ apps/bfd-db-migrator/src/test/resources/db/migration-test/error-scenarios/validation-failure/V1__Create_beneficiary_table.sql
