+++ apps/bfd-db-migrator/src/test/resources/db/migration-test/error-scenarios/validation-failure/V2__Add_column.sql
