+++ apps/bfd-model/bfd-model-rif/pom.xml
@@ -135,15 +135,6 @@
 				</exclusion>
 			</exclusions>
 		</dependency>
-
-		<dependency>
-			<!-- Provides a database schema mechanism. See the DatabaseSchemaManager 
-				class. -->
-			<groupId>org.flywaydb</groupId>
-			<artifactId>flyway-core</artifactId>
-			<version>7.8.1</version>
-		</dependency>
-
 		<dependency>
 			<!-- Used to run our unit and integration tests. -->
 			<groupId>org.junit.jupiter</groupId>
