--- apps/bfd-model/bfd-model-rif/src/test/java/gov/cms/bfd/model/rif/schema/DatabaseSchemaManagerTest.java
@@ -1,16 +1,0 @@
-package gov.cms.bfd.model.rif.schema;
-
-import org.junit.jupiter.api.Test;
-
-/** Unit tests for {@link gov.cms.bfd.model.rif.schema.DatabaseSchemaManager}. */
-public final class DatabaseSchemaManagerTest {
-  /**
-   * Verifies that {@link gov.cms.bfd.model.rif.schema.DatabaseSchemaManager} runs correctly against
-   * an HSQL database.
-   */
-  @Test
-  public void createOrUpdateSchemaOnHsql() {
-    // Ensure that this runs without errors.
-    DatabaseSchemaManager.createOrUpdateSchema(DatabaseTestUtils.get().getUnpooledDataSource());
-  }
-}
