+++ apps/bfd-pipeline/bfd-pipeline-rda-grpc-apps/pom.xml
@@ -16,9 +16,11 @@
     </description>
 
     <properties>
+        <!-- Version of gRPC related dependencies -->
+        <grpc_java.version>1.37.0</grpc_java.version>
+        <!-- Version of the plugin used to build the gRPC stubs from the Health.proto file. -->
+        <protoc.version>3.15.8</protoc.version>
         <jpa.version>1.0.2.Final</jpa.version>
-        <slf4j.version>1.7.7</slf4j.version>
-        <hikari.version>4.0.3</hikari.version>
     </properties>
 
     <dependencies>
