+++ apps/bfd-pipeline/bfd-pipeline-shared-utils/src/main/java/gov/cms/bfd/pipeline/sharedutils/PipelineTestUtils.java
@@ -2,6 +2,7 @@ package gov.cms.bfd.pipeline.sharedutils;
 
 import com.codahale.metrics.MetricRegistry;
 import com.zaxxer.hikari.HikariDataSource;
+import gov.cms.bfd.DatabaseTestUtils;
 import gov.cms.bfd.model.rda.RdaFissClaim;
 import gov.cms.bfd.model.rda.RdaFissProcCode;
 import gov.cms.bfd.model.rif.Beneficiary;
@@ -29,7 +30,7 @@ import gov.cms.bfd.model.rif.RifFilesEvent;
 import gov.cms.bfd.model.rif.SNFClaim;
 import gov.cms.bfd.model.rif.SNFClaimLine;
 import gov.cms.bfd.model.rif.SkippedRifRecord;
-import gov.cms.bfd.model.rif.schema.DatabaseTestUtils;
+import gov.cms.bfd.sharedutils.database.DatabaseSchemaManager;
 import gov.cms.bfd.sharedutils.database.DatabaseUtils;
 import gov.cms.bfd.sharedutils.exceptions.BadCodeMonkeyException;
 import java.io.InputStream;
@@ -67,7 +68,7 @@ public final class PipelineTestUtils {
 
   private static final Logger LOGGER = LoggerFactory.getLogger(PipelineTestUtils.class);
 
-  /** The singleton {@link RifLoaderTestUtils} instance to use everywhere. */
+  /** The singleton {@link PipelineTestUtils} instance to use everywhere. */
   private static PipelineTestUtils SINGLETON;
 
   /**
@@ -77,12 +78,12 @@ public final class PipelineTestUtils {
   private final PipelineApplicationState pipelineApplicationState;
 
   /**
-   * Constructs a new {@link RifLoaderTestUtils} instance. Marked <code>private</code>; use {@link
+   * Constructs a new {@link PipelineTestUtils} instance. Marked <code>private</code>; use {@link
    * #get()}, instead.
    */
   private PipelineTestUtils() {
     MetricRegistry testMetrics = new MetricRegistry();
-    DatabaseTestUtils.get().createOrUpdateSchemaForDataSource();
+    DatabaseSchemaManager.createOrUpdateSchema(DatabaseTestUtils.initUnpooledDataSource());
     this.pipelineApplicationState =
         new PipelineApplicationState(
             testMetrics,
@@ -151,7 +152,7 @@ public final class PipelineTestUtils {
       // Disable auto-commit and remember the default schema name.
       connection.setAutoCommit(false);
       Optional<String> defaultSchemaName = Optional.ofNullable(connection.getSchema());
-      if (!defaultSchemaName.isPresent()) {
+      if (defaultSchemaName.isEmpty()) {
         throw new BadCodeMonkeyException("Unable to determine default schema name.");
       }
 
@@ -161,7 +162,7 @@ public final class PipelineTestUtils {
             Optional.ofNullable(entityType.getAnnotation(Table.class));
 
         // First, make sure we found an @Table annotation.
-        if (!entityTableAnnotation.isPresent()) {
+        if (entityTableAnnotation.isEmpty()) {
           throw new BadCodeMonkeyException(
               "Unable to determine table metadata for entity: " + entityType.getCanonicalName());
         }
