--- apps/bfd-pipeline/bfd-pipeline-shared-utils/src/main/java/gov/cms/bfd/pipeline/sharedutils/databaseschema/package-info.java
@@ -1,5 +1,0 @@
-/**
- * Contains a {@link gov.cms.bfd.pipeline.sharedutils.PipelineJob} for creating/updating the
- * database schema.
- */
-package gov.cms.bfd.pipeline.sharedutils.databaseschema;
