+++ apps/bfd-shared-utils/src/test/java/gov/cms/bfd/sharedutils/database/DatabaseSchemaManagerTest.java
@@ -1,0 +1,15 @@
+package gov.cms.bfd.sharedutils.database;
+
+import gov.cms.bfd.DatabaseTestUtils;
+import org.junit.jupiter.api.Test;
+
+/** Unit tests for {@link DatabaseSchemaManager}. */
+public final class DatabaseSchemaManagerTest {
+  /** Verifies that {@link DatabaseSchemaManager} runs correctly against an HSQL database. */
+  @Test
+  public void createOrUpdateSchemaOnHsql() {
+    // Ensure that this runs without errors.
+    DatabaseSchemaManager.createOrUpdateSchema(
+        DatabaseTestUtils.getUnpooledUnmigratedHsqlDataSource());
+  }
+}
